Refatoração Database
Colunas de controle de ID todas BIGINT, para evitar erros de conversão
Server-Client.

Término envio Log de Erros para Servidor.

+++ src/etl/EtlListener.java
@@ -38,7 +38,7 @@ public class EtlListener implements ChatMessageListener {
 		try {					
 			objMensagem.setObjSessionFactory(objSessionFactory);
 			
-			objMensagem.setId(Integer.parseInt(recMessage.getId().toString()));						
+			objMensagem.setId(Long.parseLong(recMessage.getId().toString()));						
 			objMensagem.setId_sender(recMessage.getSenderId().toString());
 			objMensagem.setSender_display_name(recMessage.getSenderDisplayName().toString());			
 			objMensagem.setContent(recMessage.getContent().toString());
@@ -87,7 +87,7 @@ public class EtlListener implements ChatMessageListener {
 		try {					
 			objMensagem.setObjSessionFactory(objSessionFactory);
 			
-			objMensagem.setId(Integer.parseInt(sentMessage.getId().toString()));						
+			objMensagem.setId(Long.parseLong(sentMessage.getId().toString()));						
 			objMensagem.setId_sender(sentMessage.getSenderId().toString());
 			objMensagem.setSender_display_name(sentMessage.getSenderDisplayName().toString());			
 			objMensagem.setContent(sentMessage.getContent().toString());
