+++ src/modal/Erros_Skype_Dao.java
@@ -1,5 +1,7 @@
 package modal;
 
+import java.math.BigInteger;
+
 import org.hibernate.SQLQuery;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -15,9 +17,9 @@ public class Erros_Skype_Dao {
 	public Erros_Skype getObjErros_Skype() { return objErros_Skype; }
 	public void setObjErros_Skype(Erros_Skype objErros_Skype) { this.objErros_Skype = objErros_Skype; }
 	
-	public int getPk() {
+	public long getPk() {
 		
-		int pk = 0;
+		long pk = 0;
 		
 		final String CUSTOM_SQL = " select coalesce(max(id_geral), 0) as id_geral from erros_skype ";
 				
@@ -30,8 +32,8 @@ public class Erros_Skype_Dao {
 			qryTeste = session.createSQLQuery(CUSTOM_SQL);
 
 			for (int index = 0; index < qryTeste.list().size();) {
-				Integer objTemp = (Integer) qryTeste.list().get(index);
-				pk = objTemp.intValue();
+				BigInteger objTemp = (BigInteger) qryTeste.list().get(index);
+				pk = objTemp.longValue();
 				break;
 			}
